Added file picker for product image and added invoice PDF generation functionality for the orders

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require("../models/product");
 const { validationResult } = require('express-validator');
+const { deleteFile } = require("../utils/file-utils");
 
 exports.getAddProduct = (_, res) => {
     res.render('admin/edit-product', {
@@ -7,24 +8,29 @@ exports.getAddProduct = (_, res) => {
         path: '/admin/add-product',
         editing: false,
         hasError: false,
-        errorMessage: null
+        errorMessage: null,
+        validationErrors: []
     });
 };
 
 exports.postAddProduct = async (req, res, next) => {
     try {
-        const { title, imageUrl, description, price } = req.body;
+        const { title, description, price } = req.body;
+        const image = req.file;
         const errors = validationResult(req);
-        if (!errors.isEmpty()) {
+        if (!errors.isEmpty() || !image) {
+            const imageErrorMessage = "Attached file is not an image";
             return res.status(422).render('admin/edit-product', {
                 pageTitle: 'Add Product',
                 path: '/admin/add-product',
                 editing: false,
                 hasError: true,
-                product: { title, imageUrl, description, price },
-                errorMessage: errors.array()[0].msg
+                product: { title, description, price },
+                errorMessage: image ? errors.array()[0].msg :imageErrorMessage,
+                validationErrors: errors.array()
             });
         }
+        const imageUrl = image.path;
         const newProduct = new Product({ title, description, imageUrl, price, userId: req.user });
         await newProduct.save();
         res.redirect('/admin/products');
@@ -46,7 +52,8 @@ exports.getEditProduct = async (req, res, next) => {
             path: '/admin/edit-product',
             editing: true,
             hasError: false,
-            errorMessage: null
+            errorMessage: null,
+            validationErrors: []
         });
     } catch (error) {
         next(error);
@@ -55,16 +62,19 @@ exports.getEditProduct = async (req, res, next) => {
 
 exports.postEditProduct = async (req, res, next) => {
     try {
-        const { productId, title, imageUrl, description, price } = req.body;
+        const { productId, title, description, price } = req.body;
+        const image = req.file;
         const errors = validationResult(req);
-        if (!errors.isEmpty()) {
+        if (!errors.isEmpty() || !image) {
+            const imageErrorMessage = "Attached file is not an image";
             return res.status(422).render('admin/edit-product', {
                 pageTitle: 'Add Product',
                 path: '/admin/add-product',
                 editing: true,
                 hasError: true,
-                product: { title, imageUrl, description, price },
-                errorMessage: errors.array()[0].msg
+                product: { title, image, description, price },
+                errorMessage: image ? errors.array()[0].msg : imageErrorMessage,
+                validationErrors: errors.array()
             });
         }
         const product = await Product.findById(productId);
@@ -72,7 +82,10 @@ exports.postEditProduct = async (req, res, next) => {
             return res.redirect('/');
         }
         product.title = title;
-        product.imageUrl = imageUrl;
+        if (image) {
+            deleteFile(product.imageUrl);
+            product.imageUrl = image.path;
+        }
         product.description = description;
         product.price = price;
         await product.save();
@@ -98,6 +111,11 @@ exports.getProducts = async (req, res, next) => {
 exports.postDeleteProduct = async (req, res, next) => {
     try {
         const { productId } = req.body;
+        const product = await Product.findById(productId);
+        if (!product) {
+            next(new Error('Product Not Found'));
+        }
+        deleteFile(product.imageUrl);
         await Product.deleteOne({ _id: productId, userId: req.user._id });
         res.redirect('/admin/products');
     } catch (error) {
